+++ adapters/http-vertx-base/src/main/java/org/eclipse/hono/adapter/http/AbstractVertxBasedHttpProtocolAdapter.java
@@ -674,7 +674,8 @@ public abstract class AbstractVertxBasedHttpProtocolAdapter<T extends HttpProtoc
                             payload,
                             tenantTracker.result(),
                             tokenTracker.result(),
-                            ttd);
+                            ttd,
+                            EndpointType.EVENT.equals(endpoint) ? HttpUtils.getTimeToLive(ctx) : null);
                     customizeDownstreamMessage(downstreamMessage, ctx);
 
                     addConnectionCloseHandler(ctx, commandConsumerTracker.result(), tenant, deviceId, currentSpan);