+++ adapters/http-vertx-base/src/test/java/org/eclipse/hono/adapter/http/AbstractVertxBasedHttpProtocolAdapterTest.java
@@ -15,6 +15,7 @@ package org.eclipse.hono.adapter.http;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -448,6 +449,39 @@ public class AbstractVertxBasedHttpProtocolAdapterTest {
                 anyInt(),
                 any(MetricsTags.TtdStatus.class),
                 any());
+    }
+
+    /**
+     * Verifies that the adapter uses the time to live value set in the down stream event message.
+     */
+    @SuppressWarnings("unchecked")
+    @Test
+    public void testUploadEventWithTimeToLive() {
+        // GIVEN an adapter with a downstream event consumer attached
+        final Future<ProtonDelivery> outcome = Future.future();
+        final DownstreamSender sender = givenAnEventSenderForOutcome(outcome);
+
+        final HttpServer server = getHttpServer(false);
+        final AbstractVertxBasedHttpProtocolAdapter<HttpProtocolAdapterProperties> adapter = getAdapter(server, null);
+
+        // WHEN a device publishes an event with a time to live value as a header
+        final Buffer payload = Buffer.buffer("some payload");
+        final HttpServerResponse response = mock(HttpServerResponse.class);
+        final HttpServerRequest request = mock(HttpServerRequest.class);
+        when(request.getHeader(eq(Constants.HEADER_TIME_TO_LIVE))).thenReturn("10");
+        final RoutingContext ctx = newRoutingContext(payload, "application/text", request, response);
+        when(ctx.addBodyEndHandler(any(Handler.class))).thenAnswer(invocation -> {
+            final Handler<Void> handler = invocation.getArgument(0);
+            handler.handle(null);
+            return 0;
+        });
+
+        adapter.uploadEventMessage(ctx, "tenant", "device");
+
+        // verifies that the downstream message contains the time to live value
+        final ArgumentCaptor<Message> messageCaptor = ArgumentCaptor.forClass(Message.class);
+        verify(sender).sendAndWaitForOutcome(messageCaptor.capture(), any());
+        assertEquals(10L * 1000, messageCaptor.getValue().getTtl());
     }
 
     /**